feat(story-list): add story item card xml

+++ app/src/main/java/com/dicoding/android/intermediate/submission/storyapp/models/responses/StoryListResponse.kt
@@ -32,11 +32,11 @@ data class StoryListItem(
 	val description: String? = null,
 
 	@field:SerializedName("lon")
-	val lon: Double? = null,
+	val lon: Float? = null,
 
 	@field:SerializedName("id")
 	val id: String? = null,
 
 	@field:SerializedName("lat")
-	val lat: Double? = null
+	val lat: Float? = null
 ) : Parcelable
